+++ src/Tooltip.cpp
@@ -605,7 +605,7 @@ void __fastcall Tooltip_ShowItemInfo_2(int slot)
 		for( int i = 0; i < countof(item.effect); ++i ){
 			Effect& e = item.effect[i];
 			if( e.id == AE_NO || e.origin >= AO_BASE ) continue;
-			if( item.MagicLevel >= ML_2_UNIQUE && (e.minVal < 0 || e.chance < 0) && IsUniquePenaltyEffect(e.id) ) continue;
+			if( item.MagicLevel >= ML_2_UNIQUE && (e.minVal < 0 || e.chance < 0) && IsHarmfulWhenNegative(e.id) ) continue;
 			if( item.MagicLevel >= ML_2_UNIQUE && e.id == AE_MINUS_LIGHT_RADIUS ) continue;
 			if( PrintItemEffectToStringBuffer(e.id, &item, i) ){
 				Tooltip_AddLine_2(InfoPanelBuffer, C_1_Blue);
@@ -1447,7 +1447,7 @@ void __fastcall Tooltip_ShowItemInfo_3(int slot)
 		for( int i = 0; i < countof(item.effect); ++i ){
 			Effect& e = item.effect[i];
 			if( e.id == AE_NO || e.origin >= AO_BASE ) continue;
-			if( item.MagicLevel >= ML_2_UNIQUE && (e.minVal < 0 || e.chance < 0) && IsUniquePenaltyEffect(e.id) ) continue;
+			if( item.MagicLevel >= ML_2_UNIQUE && (e.minVal < 0 || e.chance < 0) && IsHarmfulWhenNegative(e.id) ) continue;
 			if( item.MagicLevel >= ML_2_UNIQUE && e.id == AE_MINUS_LIGHT_RADIUS ) continue;
 			if( PrintItemEffectToStringBuffer(e.id, &item, i) ){
 				Tooltip_AddLine_3(InfoPanelBuffer, C_1_Blue);
@@ -2397,7 +2397,7 @@ void __fastcall Tooltip_ShowItemInfo( int slot )
 		for( int i = 0; i < countof(item.effect); ++i ){
 			Effect& e = item.effect[i];
 			if( e.id == AE_NO || e.origin >= AO_BASE ) continue;
-			if( item.MagicLevel >= ML_2_UNIQUE && (e.minVal < 0 || e.chance < 0) && IsUniquePenaltyEffect(e.id) ) continue;
+			if( item.MagicLevel >= ML_2_UNIQUE && (e.minVal < 0 || e.chance < 0) && IsHarmfulWhenNegative(e.id) ) continue;
 			if( item.MagicLevel >= ML_2_UNIQUE && e.id == AE_MINUS_LIGHT_RADIUS ) continue;
 			if( PrintItemEffectToStringBuffer(e.id, &item, i) ){
 				Tooltip_AddLine(InfoPanelBuffer, C_1_Blue);